refactor: modularize services
