+++ FixChain/modules/fix/__init__.py
@@ -1,0 +1,8 @@
+from .registry import register, create
+from .llm import LLMFixer
+
+__all__ = [
+    "register",
+    "create",
+    "LLMFixer",
+]
