+++ FixChain/modules/rag/__init__.py
@@ -1,0 +1,8 @@
+from .registry import register, create
+from .mongodb import MongoDBRAG
+
+__all__ = [
+    "register",
+    "create",
+    "MongoDBRAG",
+]
